chore: add Header component to MainLayout
