+++ frontend/src/index.css
@@ -1,7 +1,5 @@
 @import "tailwindcss";
 
 #root {
-  max-width: 1280px;
-  margin: 0 auto;
   text-align: center;
 }
